requirements.txt
cycler==0.11.0
fonttools==4.37.1
kiwisolver==1.4.4
matplotlib==3.5.3
nidaqmx==0.6.4
numpy==1.23.2
packaging==21.3
Pillow==9.2.0
psycopg2==2.9.3
pyparsing==3.0.9
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
pyqtgraph==0.12.4
python-dateutil==2.8.2
PyYAML==6.0
six==1.16.0
tzdata==2022.2
